# Optimize uint32writer