Don't use `binary.Write` which makes 2 allocs. Instead since we know we always want a uint32 encode this directly with `binary.PutUint32`, which is what `binary.Write` does.

Creates a fast path and a slow path. Fast path is for types that satisfy `marshaller` where we can work with just a single buffer.
Slow path is for everything else where it uses `proto.Marshal`.

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkUint32DelimWriterMarshaller-8     159           77.8          -51.07%
BenchmarkUint32DelimWriterFallback-8       879           764           -13.08%

benchmark                                  old allocs     new allocs     delta
BenchmarkUint32DelimWriterMarshaller-8     3              0              -100.00%
BenchmarkUint32DelimWriterFallback-8       9              7              -22.22%

benchmark                                  old bytes     new bytes     delta
BenchmarkUint32DelimWriterMarshaller-8     80            0             -100.00%
BenchmarkUint32DelimWriterFallback-8       728           712           -2.20%
```