@awalterschulze 2017-03-30
Couldn't the slow stream problem also be solved by the user with the slow stream passing in a buffered writer into the NewDelimitedWriter?