@awalterschulze 2017-03-30
Good catch at binary.Write :)