@cpuguy83 2017-03-30
@awalterschulze Yes, the caller could pass in a buffer but I don't think this would improve anything since the writes are opaque at this point.