@awalterschulze 2017-04-01
Can we try some bigger structs in the benchmark.
I still can't believe that copying the buffer of the message twice instead of once is faster.

So what you mean with opaque is that the bufferedwriter would not be able to tell which writes are length writes and which are marshaled buffer writes and so not be able to deduce the most efficient logic?