@cpuguy83 2017-04-05
It would be nice to have a bigger proto that implements `MarshalTo` since there's really two paths here.
But in any case, changed this around a little bit so it doesn't even need the 2nd buffer.
Now if the past in proto meets all of these, it will have no allocations:

- Implements `MarshalTo()`
- `getSize()` works
- Passed in buffer is smaller than the internal buffer

This is the fast path.
If the passed in proto is larger than the internal buffer than it will require an alloc.

The slow path's allocations is reduced by 2.
Only 1 allocation is required for the slow path except for anything that is incurred by the actual marshaller.

Not a huge fan of how this can store larger and larger buffers forever, but that is the existing behavior.

See the original comment for updated benchmark results.