@awalterschulze 2017-04-05
All the thetest.proto structs with a MarshalTo are available in ./test/combos/both