@awalterschulze 2017-04-05
It would be nice to compare the same struct for fallback vs non fallback.