@awalterschulze 2017-04-05
The uint32Reader has a maxSize to avoid the problem of the buffer growing ever larger.
But maybe lets tackle that in a separate commit, because it will change the api or at least add another extra constructor.