@cpuguy83 2017-04-05
@awalterschulze Can you explain how I can use these without causing panics?

```
<truncated>
2017/04/05 08:42:51 proto: duplicate proto type registered: test.NinOptNonByteCustomType
2017/04/05 08:42:51 proto: duplicate proto type registered: test.NidRepNonByteCustomType
2017/04/05 08:42:51 proto: duplicate proto type registered: test.NinRepNonByteCustomType
2017/04/05 08:42:51 proto: duplicate proto type registered: test.ProtoType
panic: proto: duplicate enum registered: test.TheTestEnum

goroutine 1 [running]:
github.com/gogo/protobuf/proto.RegisterEnum(0x1328308, 0x10, 0xc420019200, 0xc420019230)
        /Users/bgoff/go/src/github.com/gogo/protobuf/proto/properties.go:907 +0x2ac
github.com/gogo/protobuf/test/combos/both.init.1()
        /Users/bgoff/go/src/github.com/gogo/protobuf/test/combos/both/thetest.pb.go:1505 +0xc8a
github.com/gogo/protobuf/test/combos/both.init()
        /Users/bgoff/go/src/github.com/gogo/protobuf/test/combos/both/uuid.go:134 +0x88f
github.com/gogo/protobuf/io_test.init()
        /Users/bgoff/go/src/github.com/gogo/protobuf/io/uint32_test.go:69 +0x76
main.init()
        github.com/gogo/protobuf/io/_test/_testmain.go:66 +0x58
```