@awalterschulze 2017-04-05
Ah damn, I forgot about stupid Register functions. Lets leave the benchmarking of the same struct :(
I really want to make a plugin that just does not generate the code to Register any of the messages.