@cpuguy83 2017-04-05
There seems to be something breaking here not covered by tests. I'll look into it.