@cpuguy83 2017-04-05
Went ahead and changed the ordering so `PutUint32` is before writing the proto to the buffer.