@awalterschulze 2017-04-05
The rest looks good :)
Also would you like to add yourself as an AUTHOR or CONTRIBUTOR whichever is appropriate?