@awalterschulze 2017-04-05
If checks pass I'll merge :)