@awalterschulze 2017-04-05
Thank you very much :)  Really cool optimization.