@cpuguy83 2017-04-05
Thanks, sorry for the back and forth there!