@awalterschulze 2017-04-05
No problem that's how it usually goes ☺️

On Wed, 5 Apr 2017, 19:15 Brian Goff, <notifications@github.com> wrote:

> Thanks, sorry for the back and forth there!
>
> —
> You are receiving this because you modified the open/close state.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/gogo/protobuf/pull/281#issuecomment-291931865>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABvsLdA4gVnn2HNVYjCCQn5Do_hl4e1eks5rs8wtgaJpZM4Mtl3w>
> .
>